Harden URL filtering, split caches, fix scroll race
Summary: Fix 15 failing URL normalization tests, resolve infinite scroll
race condition on back-navigation, split monolithic cache classes, and
move test infrastructure to the projects that own it.

Implementation:
• Move SectionCache model to TechHub.Core so Web and tests can reference
  it without circular dependencies
• Split SectionCache.cs into SectionCache (Core), SectionCacheRefreshService,
  and SectionCacheHealthCheck; same split applied to HeroBannerCache
• Fix UrlNormalizationMiddleware: move !IsReady guard from IsLegacyLookupCandidate
  to InvokeAsync; add IsReady guard in TryRedirectLegacyFeed
• Fix scroll-manager.js race: store pendingIntersect flag when IO fires during
  back-nav traversal; flush it in finishNavigation after navigating = false
• Add SectionCacheBuilder and A.SectionCache to TestUtilities/Builders
• Move TechHubApiFactory from TestUtilities → Api.Tests (only used there)
• Move DatabaseFixture from TestUtilities → Infrastructure.Tests (only used there)
• Slim down TechHub.TestUtilities: remove TechHub.Api ref, Mvc.Testing, Moq,
  Npgsql, Testcontainers packages; add them directly to Api.Tests / Infrastructure.Tests
• Extract AdminRequests.cs from AdminEndpoints.cs
• Add EnsureSectionTagsTests and HtmlToMarkdownConverterTests
• Update AGENTS.md files for Api.Tests, Infrastructure.Tests, TestUtilities
