+++ src/TechHub.Core/Models/Core/SectionCache.cs
@@ -1,11 +1,9 @@
-using TechHub.Core.Models;
-
-namespace TechHub.Web.Services;
+namespace TechHub.Core.Models;
 
 /// <summary>
 /// In-memory cache of sections loaded at application startup.
 /// Provides synchronous access to section data for immediate rendering without flicker.
-/// Periodically refreshes from the API to stay current after deployments.
+/// Periodically refreshed from the API by <c>SectionCacheRefreshService</c>.
 /// </summary>
 public class SectionCache
 {
@@ -14,7 +12,7 @@ public class SectionCache
     /// with a single volatile reference swap. Readers get a consistent snapshot
     /// without any locking: they read <c>_data</c> once and use that snapshot throughout.
     /// </summary>
-    private record CacheData(
+    private sealed record CacheData(
         Dictionary<string, Section> ByName,
         Dictionary<string, HashSet<string>> BySection,
         HashSet<string> AllCollections,
@@ -84,65 +82,5 @@ public class SectionCache
     public bool IsKnownCollectionInAnySection(string collectionName)
     {
         return Volatile.Read(ref _data).AllCollections.Contains(collectionName);
-    }
-}
-
-/// <summary>
-/// Background service that periodically refreshes the SectionCache from the API.
-/// Prevents stale navigation data when the API is redeployed with new configuration.
-/// </summary>
-public class SectionCacheRefreshService : BackgroundService
-{
-    private static readonly TimeSpan _refreshInterval = TimeSpan.FromMinutes(5);
-
-    private readonly IServiceScopeFactory _scopeFactory;
-    private readonly SectionCache _sectionCache;
-    private readonly ILogger<SectionCacheRefreshService> _logger;
-
-    public SectionCacheRefreshService(
-        IServiceScopeFactory scopeFactory,
-        SectionCache sectionCache,
-        ILogger<SectionCacheRefreshService> logger)
-    {
-        _scopeFactory = scopeFactory;
-        _sectionCache = sectionCache;
-        _logger = logger;
-    }
-
-    protected override async Task ExecuteAsync(CancellationToken stoppingToken)
-    {
-        while (!stoppingToken.IsCancellationRequested)
-        {
-            await Task.Delay(_refreshInterval, stoppingToken);
-
-            try
-            {
-                using var scope = _scopeFactory.CreateScope();
-                var apiClient = scope.ServiceProvider.GetRequiredService<TechHubApiClient>();
-                var sections = await apiClient.GetAllSectionsAsync(stoppingToken);
-
-                if (sections != null)
-                {
-                    var sectionList = sections.ToList();
-                    if (sectionList.Count > 0)
-                    {
-                        _sectionCache.Initialize(sectionList);
-                        _logger.LogDebug("SectionCache refreshed with {Count} sections", sectionList.Count);
-                    }
-                }
-            }
-            catch (OperationCanceledException) when (stoppingToken.IsCancellationRequested)
-            {
-                // Normal shutdown
-            }
-            catch (HttpRequestException ex)
-            {
-                _logger.LogWarning(ex, "Failed to refresh SectionCache, will retry in {Interval}", _refreshInterval);
-            }
-            catch (TaskCanceledException ex) when (!stoppingToken.IsCancellationRequested)
-            {
-                _logger.LogWarning(ex, "SectionCache refresh timed out, will retry in {Interval}", _refreshInterval);
-            }
-        }
     }
 }
