+++ src/TechHub.Web/Components/Layout/MainLayout.razor
@@ -1,4 +1,5 @@
 ﻿@inherits LayoutComponentBase
+@using TechHub.Core.Models
 @using TechHub.Web.Services
 @inject NavigationManager Navigation
 @inject SectionCache SectionCache
