+++ src/TechHub.Web/Components/Pages/Admin/AdminSettings.razor
@@ -2,6 +2,7 @@
 @layout AdminLayout
 @rendermode InteractiveServer
 @using TechHub.Core.Models.Admin
+@using TechHub.Core.Models
 @using TechHub.Web.Services
 @implements IDisposable
 @inject ITechHubApiClient Api
