+++ tests/TechHub.Infrastructure.Tests/Repositories/ProcessedUrlRepositoryTests.cs
@@ -2,7 +2,6 @@ using FluentAssertions;
 using Microsoft.Extensions.Logging.Abstractions;
 using TechHub.Infrastructure.Data;
 using TechHub.Infrastructure.Repositories;
-using TechHub.TestUtilities;
 
 namespace TechHub.Infrastructure.Tests.Repositories;
