+++ tests/TechHub.Infrastructure.Tests/Repositories/SectionRoundupRepositoryTests.cs
@@ -2,7 +2,6 @@ using Dapper;
 using FluentAssertions;
 using Microsoft.Extensions.Logging.Abstractions;
 using TechHub.Infrastructure.Repositories;
-using TechHub.TestUtilities;
 
 namespace TechHub.Infrastructure.Tests.Repositories;
