+++ tests/TechHub.Infrastructure.Tests/TechHub.Infrastructure.Tests.csproj
@@ -13,6 +13,8 @@
     <PackageReference Include="Microsoft.Extensions.Caching.Memory" />
     <PackageReference Include="Microsoft.Testing.Extensions.CodeCoverage" />
     <PackageReference Include="Moq" />
+    <PackageReference Include="Npgsql" />
+    <PackageReference Include="Testcontainers.PostgreSql" />
     <PackageReference Include="xunit.v3.mtp-v2" />
     <PackageReference Include="YoutubeExplode" />
   </ItemGroup>
